# Optimized Recommeder(lsh, euclid_lsh)'s random projection