--- a/jubatus/core/nearest_neighbor/lsh.cpp
+++ b/jubatus/core/nearest_neighbor/lsh.cpp
@@ -55,16 +55,7 @@ void lsh::set_config(const config& conf) {
         common::invalid_parameter("1 <= hash_num"));
   }
   threads_ = read_threads_config(conf.threads);
-
-  if (conf.cache_size.bool_test()) {
-    if (!(0 <= *(conf.cache_size))) {
-      throw JUBATUS_EXCEPTION(
-          common::invalid_parameter("0 <= cache_size"));
-    }
-    if (*(conf.cache_size) > 0) {
-      cache_.reset(new random_projection_cache(*(conf.cache_size)));
-    }
-  }
+  init_cache_from_config(cache_, conf.cache_size);
 }
 
 }  // namespace nearest_neighbor