--- a/jubatus/core/recommender/euclid_lsh.cpp
+++ b/jubatus/core/recommender/euclid_lsh.cpp
@@ -33,13 +33,13 @@
 #include "../storage/lsh_vector.hpp"
 #include "../storage/lsh_index_storage.hpp"
 #include "../unlearner/unlearner_factory.hpp"
+#include "../nearest_neighbor/bit_vector_ranking.hpp"
 
 using std::string;
 using std::vector;
 using std::pair;
 using std::ostream;
 using std::istream;
-using jubatus::util::math::random::mtrand;
 using jubatus::util::concurrent::scoped_lock;
 
 namespace jubatus {
@@ -64,14 +64,6 @@ float calc_norm(const common::sfv_t& sfv) {
   return std::sqrt(sqnorm);
 }
 
-void calc_projection(uint32_t seed, size_t size, vector<float>& ret) {
-  mtrand rnd(seed);
-  ret.resize(size);
-  for (size_t i = 0; i < size; ++i) {
-    ret[i] = rnd.next_gaussian();
-  }
-}
-
 }  // namespace
 
 euclid_lsh::config::config()
@@ -79,7 +71,8 @@ euclid_lsh::config::config()
       table_num(DEFAULT_TABLE_NUM),
       bin_width(DEFAULT_BIN_WIDTH),
       probe_num(DEFAULT_NUM_PROBE),
-      seed(DEFAULT_SEED) {
+      seed(DEFAULT_SEED),
+      threads(), cache_size() {
 }
 
 const uint64_t euclid_lsh::DEFAULT_HASH_NUM = 64;  // should be in config
@@ -97,13 +90,14 @@ euclid_lsh::euclid_lsh()
                                         DEFAULT_TABLE_NUM,
                                         DEFAULT_SEED)))),
       bin_width_(DEFAULT_BIN_WIDTH),
-      num_probe_(DEFAULT_NUM_PROBE) {
+      num_probe_(DEFAULT_NUM_PROBE), threads_(1), cache_() {
 }
 
 euclid_lsh::euclid_lsh(const config& config)
     : mixable_storage_(),
       bin_width_(config.bin_width),
-      num_probe_(config.probe_num) {
+      num_probe_(config.probe_num),
+      threads_(nearest_neighbor::read_threads_config(config.threads)) {
 
   if (!(1 <= config.hash_num)) {
     throw JUBATUS_EXCEPTION(
@@ -129,6 +123,8 @@ euclid_lsh::euclid_lsh(const config& config)
     throw JUBATUS_EXCEPTION(
         common::invalid_parameter("0 <= seed"));
   }
+
+  nearest_neighbor::init_cache_from_config(cache_, config.cache_size);
 
   typedef storage::mixable_lsh_index_storage mli_storage;
   typedef mli_storage::model_ptr model_ptr;
@@ -253,24 +249,12 @@ framework::mixable* euclid_lsh::get_mixable() const {
 }
 
 vector<float> euclid_lsh::calculate_lsh(const common::sfv_t& query) const {
-  vector<float> hash(mixable_storage_->get_model()->all_lsh_num());
-  for (size_t i = 0; i < query.size(); ++i) {
-    const uint32_t seed = common::hash_util::calc_string_hash(query[i].first);
-    const vector<float> proj = get_projection(seed);
-    for (size_t j = 0; j < hash.size(); ++j) {
-      hash[j] += query[i].second * proj[j];
-    }
-  }
+  vector<float> hash = nearest_neighbor::random_projection(
+    query, mixable_storage_->get_model()->all_lsh_num(), threads_, cache_);
   for (size_t j = 0; j < hash.size(); ++j) {
     hash[j] /= bin_width_;
   }
   return hash;
-}
-
-vector<float> euclid_lsh::get_projection(uint32_t seed) const {
-  vector<float> proj;
-  calc_projection(seed, mixable_storage_->get_model()->all_lsh_num(), proj);
-  return proj;
 }
 
 void euclid_lsh::initialize_model() {