--- a/jubatus/core/recommender/lsh_util.cpp
+++ b/jubatus/core/recommender/lsh_util.cpp
@@ -21,6 +21,7 @@
 #include <vector>
 #include "jubatus/util/math/random.h"
 #include "../common/assert.hpp"
+#include "../common/hash.hpp"
 #include "../storage/bit_vector.hpp"
 
 namespace jubatus {
@@ -53,22 +54,16 @@ void set_bit_vector(const std::vector<float>& vec, bit_vector& bit_vec) {
 }
 
 void prod_invert_and_vector(
-    const unordered_map<string, vector<float> >& matrix,
     const common::sfv_t& vec,
     size_t dim,
     vector<float>& ret) {
   vector<float> r(dim);
   for (size_t i = 0; i < vec.size(); ++i) {
     const string& column = vec[i].first;
+    const uint32_t seed = common::hash_util::calc_string_hash(column);
+    vector<float> v;
+    generate_random_vector(dim, seed, v);
     float val = vec[i].second;
-    unordered_map<string, vector<float> >::const_iterator it =
-        matrix.find(column);
-    if (it == matrix.end()) {
-      continue;
-    }
-    const vector<float>& v = it->second;
-
-    JUBATUS_ASSERT_EQ(v.size(), r.size(), "");
     for (size_t j = 0; j < v.size(); ++j) {
       r[j] += v[j] * val;
     }