--- a/jubatus/core/recommender/lsh_util.hpp
+++ b/jubatus/core/recommender/lsh_util.hpp
@@ -38,8 +38,6 @@ void set_bit_vector(
     jubatus::core::storage::bit_vector& bit_vec);
 
 void prod_invert_and_vector(
-    const jubatus::util::data::unordered_map<std::string, std::vector<float> >&
-        matrix,
     const common::sfv_t& vec,
     size_t dim,
     std::vector<float>& ret);