--- a/jubatus/core/recommender/wscript
+++ b/jubatus/core/recommender/wscript
@@ -14,7 +14,6 @@ def build(bld):
     'minhash.cpp',
     'lsh.cpp',
     'recommender_factory.cpp',
-    'lsh_util.cpp',
     'euclid_lsh.cpp',
     'nearest_neighbor_recommender.cpp',
     ]
@@ -42,7 +41,6 @@ def build(bld):
       'recommender_mock_storage_test.cpp',
       'recommender_mock_test.cpp',
       'recommender_random_test.cpp',
-      'lsh_util_test.cpp',
       'euclid_lsh_test.cpp',
       'minhash_test.cpp',
       'lsh_test.cpp',