This pull request is based from non-merged branch.
based branch: #247 

NNと同様にLSH計算のSIMD化・並列化，キャッシュ機能のLRU化を行ないました．

Recommenderのlshは常にキャッシュ機能が有効だったため，
非常に多くのメモリを消費していました．
euclid_lshはオプションでキャッシュの有効無効を
設定できましたが，有効か無効かしか設定できず，
文章などの冪分布ではメモリ消費効率がよく有りませんでしたし
無効にすると性能が悪化していました．

このプルリクエストではNNと同様の高速化を盛り込むことで
文章等ではキャッシュを利用するよりも二倍程度高速でありながら
メモリ消費量を大きく削減しています．

```
Benchmark Results
(Recommender, lsh, hash_num=8192, haswell@2.4GHz)

           update_row     memory
20news:    latency[ms]  usage[MiB]
  baseline:    17.8829  3934.240
  SIMD:         7.4233   391.125
  cache=8192:   4.4869   648.363
  thread=8:     1.7390   391.173
  cache+thread: 2.9309   648.433

iris:
  baseline:   0.772004   4.20097
  SIMD:       0.246876   1.36974
```

スレッドとキャッシュを両方組み合わせると性能が低下しますが，
LRUのロック競合によるものと思われます．
(LRUなのでgetするときもライトロックが必要なため，mutexでLRUへのアクセスを保護しています)