@kazuki 2016-04-27
rebased non-merged branch #247